Fixed Pipeline Warnings
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/loopsummary@37345 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/loopsummary/strategies/extrapolation/ConstantExtrapolationStrategy.java
@@ -231,7 +231,8 @@ public class ConstantExtrapolationStrategy extends AbstractExtrapolationStrategy
                 expressionType,
                 calculationType,
                 loopBound,
-                CIntegerLiteralExpression.createDummyLiteral(2 * boundDelta, CNumericTypes.INT),
+                CIntegerLiteralExpression.createDummyLiteral(
+                    2 * ((long) boundDelta), CNumericTypes.INT),
                 BinaryOperator.MINUS),
             BinaryOperator.LESS_THAN);
     CFAEdge twiceLoopUnrollingConditionEdgeTrue =
